+++ CHANGELOG.md
@@ -8,6 +8,7 @@ e [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Added
 - Implementado controle de horário (final e inicial) em OS. [@Pr3d4dor](https://github.com/Pr3d4dor)
+- Adequação para exibir valor de total de OS em home. [@Pr3d4dor](https://github.com/Pr3d4dor)
 
 ## [4.11.1] - 2020-07-18